merge code from develop

+++ Timer_Rubik.WebApp/Models/Favorite.cs
@@ -8,7 +8,7 @@
 
         public Guid ScrambleId { get; set; }
 
-        public long Time { get; set; }
+        public float Time { get; set; }
 
         public DateTime CreatedAt { get; set; }
